Fix handling of offline users
This didn't cause problems earlier because the only anonymous users we saw were
web clients users, and apparently the server was not notified when they went
offline. However the HTTP API does a "real" XMPP connection, not BOSH, so when
it is killed the server now gets an unavailable presence, makeing it crash.

Thanks to Justin for pointing that out.