Change the way the test server emits IQ stanzas
Not specifying the IQ type in the event name actually makes things easier. It is
now required to check for that type in tests, but this means that if there's an
incoming stanza with an incorrect type (i.e. expected "error" but got "result"),
this will be detected instantly instead of triggering a test timeout.