Add a giant MAM test
Only one huge test for now, but it actually tests a lot of things: MAM messages
for posts, subscriptions, affiliations and configurations. Also checks that they
are only delivered to the right user, and that data relevant only to other users
are not included.

With this commit several tests fail, and for very good reasons.