Workaround callback woes
That was a tricky one.

The constructor of the Request class takes a callback, which is called
in the constructor. And the PubSub backend creates instances of the
Request class like this:

    req = new Request ->
       localPushData = req.localPushData?()

But here's the thing: if the callback is called directly from the
constructor (without being delayed due to IO), req will be accessed
before the constructor has returned, therefore req will be "undefined",
causing this very nice error:

    TypeError: Cannot read property 'localPushData' of undefined

In the real world this doesn't happen because the Request constructor
communicates with the XMPP server, causing the callback to be called
later in the event loop. But in the test server, these communications
are instantaneous, so there's no delay before calling the callback.

This shouldn't change anything in real use cases.